requirements.txt
# ============================================
# Core Dependencies
# ============================================
streamlit==1.31.0
boto3==1.34.0
pyyaml==6.0.1
numpy==1.26.3

# ============================================
# Document Processing
# ============================================
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ============================================
# LangChain & RAG
# ============================================
langchain==0.1.0
langchain-aws==0.1.0
langchain-community==0.0.20
langchain-core==0.1.10
langchain-text-splitters==0.0.1

# ============================================
# Agentic RAG (LangGraph)
# ============================================
langgraph>=0.2.0,<1.0.0

# ============================================
# Web Search Fallback
# ============================================
tavily-python>=0.3.0

# ============================================
# Vector Store
# ============================================
faiss-cpu==1.7.4

# ============================================
# Testing (Optional)
# ============================================
pytest==7.4.3
pytest-cov==4.1.0

# ============================================
# Development (Optional)
# ============================================
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0

# ============================================
# Notes
# ============================================
# - All versions are pinned for reproducibility
# - Use 'pip install -r requirements.txt' to install
# - For development: pip install -r requirements.txt -r requirements-dev.txt
# - Tested with Python 3.8+
